refactor: rewrite api to use a config file for referrers
Closes #1
