+++ .gitignore
@@ -12,3 +12,4 @@ myenv/
 ENV/ 
 env.bak/ 
 venv.bak/
+.env*.local
