+++ config.yml
@@ -1,0 +1,3 @@
+referrers:
+  portfolio: "https://minch.dev/contact"
+  test: "/form"
