+++ requirements.txt
@@ -1,2 +1,3 @@
 Flask
 psycopg2-binary
+PyYAML
